src/backend/youtube_scrapper_for_internship-main/requirements.txt
# Flask backend
flask
flask-cors
flask-sqlalchemy
qrcode
pillow

# YouTube scraper
google-api-python-client==2.121.0
python-dotenv